+++ lib/ff_annotation_route_library.dart
@@ -2,8 +2,15 @@ library ff_annotation_route_library;
 
 export 'package:ff_annotation_route_core/ff_annotation_route_core.dart';
 
+export 'src/global_navigator.dart';
 export 'src/helper.dart';
+export 'src/interceptor/extension.dart';
+export 'src/interceptor/navigator_with_interceptor.dart';
+export 'src/interceptor/route_interceptor.dart';
 export 'src/page.dart';
 export 'src/route_helper.dart';
 export 'src/route_information_parser.dart';
+export 'src/route_lifecycle/extended_route_aware.dart';
+export 'src/route_lifecycle/extended_route_observer.dart';
+export 'src/route_lifecycle/route_lifecycle_state.dart';
 export 'src/router_delegate.dart';